chore(types): Extract pagination types to separate module
