+++ packages/types/src/index.retheme.ts
@@ -53,3 +53,4 @@ export * from './verification';
 export * from './web3';
 export * from './web3Wallet';
 export * from './customPages';
+export * from './pagination';
